requirements.txt
contourpy==1.3.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.54.1
imageio==2.36.0
kiwisolver==1.4.7
lazy_loader==0.4
mahotas==1.4.18
matplotlib==3.9.2
networkx==3.4.1
numpy==2.1.1
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pyfeats==0.0.11
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.7.0
scikit-image==0.24.0
scipy==1.14.1
six==1.16.0
tifffile==2024.9.20
tzdata==2024.2
